Created example map for reference
Added a simple mesh that shows how maps should be created.

NOTE:
The base mesh is what will be displayed in play. The meshes with the prefix "UW" and "WA" are unwalkable and walkable surfaces respectively. These meshes will not be displayed during normal play and are only used for generating the nav mesh.